# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SeuDrink API",
        "description": "Documentação da API usando Swagger",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/api/drink": {
            "get": {
                "tags": [
                    "Drinks"
                ],
                "summary": "Lista todos os drinks filtrados",
                "operationId": "29a7e7fc56b6f4cd34f20fb7a64d5417",
                "parameters": [
                    {
                        "name": "fruta_id",
                        "in": "query",
                        "description": "ID da fruta usada no drink",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "bebida_id",
                        "in": "query",
                        "description": "ID da bebida usada no drink",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de drinks retornada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Drink"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drink/novo": {
            "post": {
                "tags": [
                    "Drinks"
                ],
                "summary": "Cria um novo drink",
                "operationId": "0a22e9faec73f45a305c1c8aeec28ed2",
                "parameters": [
                    {
                        "name": "ingrediente_ids[]",
                        "in": "query",
                        "description": "IDs dos ingredientes",
                        "required": true,
                        "schema": {
                            "description": "Lista de IDs dos ingredientes",
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "foto",
                                    "preparo",
                                    "fruta_id",
                                    "bebida_id"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Mojito"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "preparo": {
                                        "type": "string",
                                        "example": "Misture hortelã, limão e rum"
                                    },
                                    "fruta_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bebida_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Drink criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Drink"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/drink/{id}": {
            "post": {
                "tags": [
                    "Drinks"
                ],
                "summary": "Atualiza um drink existente",
                "operationId": "bd1b2840677921c24fca3f981a73adc0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do drink a ser atualizado",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "ingrediente_ids[]",
                        "in": "query",
                        "description": "IDs dos ingredientes",
                        "required": true,
                        "schema": {
                            "description": "Lista de IDs dos ingredientes",
                            "type": "array",
                            "items": {
                                "type": "integer"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "foto",
                                    "preparo",
                                    "fruta_id",
                                    "bebida_id"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Mojito"
                                    },
                                    "foto": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "preparo": {
                                        "type": "string",
                                        "example": "Misture hortelã, limão e rum"
                                    },
                                    "fruta_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "bebida_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Drink atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Drink"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Drink não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Drink": {
                "title": "Drink",
                "description": "Modelo de um Drink",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nome": {
                        "type": "string",
                        "example": "Caipirinha"
                    },
                    "foto": {
                        "type": "string",
                        "example": "https://example.com/drink.jpg"
                    },
                    "preparo": {
                        "type": "string",
                        "example": "Misture todos os ingredientes e sirva."
                    },
                    "fruta_id": {
                        "type": "integer",
                        "example": 2
                    },
                    "bebida_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}